Refactor: Refactor server side folder structure
