--- be/src/allowance/index.ts
@@ -1,21 +1,0 @@
-import { Router } from "express";
-import { verifyToken } from "../middlewares/verifyToken";
-import {
-  deleteAllowance,
-  getAllowance,
-  postAllowance,
-  updateAllowance,
-} from "./controller";
-
-const router = Router();
-
-const routes = () => {
-  router.post("/v1/allowance", verifyToken, postAllowance);
-  router.get("/v1/allowance", verifyToken, getAllowance);
-  router.delete("/v1/allowance/:id", verifyToken, deleteAllowance);
-  router.put("/v1/allowance/:id", verifyToken, updateAllowance);
-  // router.get("/v1/users", getUser);
-  return router;
-};
-
-export default routes;
