+++ be/src/api/v1/interfaces/types/position.types.ts
@@ -1,0 +1,12 @@
+export type TPosition = {
+  position_cd: string;
+  position_desc: string;
+  position_desc_nep?: string;
+  upper_position_cd?: string;
+  group_flag: string;
+  disabled: string;
+  entered_by: string;
+  order_no: number;
+  computer_position_cd?: string;
+  new_position_cd?: string;
+};
